Move/fix tests. Tweak names of TurbolinksNavRule
