+++ turbolinks/src/main/kotlin/com/basecamp/turbolinks/config/PathConfiguration.kt
@@ -5,8 +5,8 @@ import android.content.Context
 import android.net.Uri
 import com.basecamp.turbolinks.BuildConfig
 import com.basecamp.turbolinks.util.TurbolinksLog
-import com.basecamp.turbolinks.nav.TurbolinksNavigationRule.Presentation
-import com.basecamp.turbolinks.nav.TurbolinksNavigationRule.PresentationContext
+import com.basecamp.turbolinks.nav.TurbolinksNavRule.Presentation
+import com.basecamp.turbolinks.nav.TurbolinksNavRule.PresentationContext
 import com.google.gson.annotations.SerializedName
 import java.net.URL
 import java.util.regex.PatternSyntaxException
